pyproject.toml
[tool.poetry]
name = "niimprint"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "~3.11"
pyserial = "^3.5"
pillow = "^10.1.0"
click = "^8.1.7"
click-option-group = "^0.5.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["I", "F", "E"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
